ci: fix redis-cluster config for macos, remove unneeded docker env.sh
