+++ README.md
@@ -191,7 +191,6 @@ If you want to run integration tests in a repetitive manner, you can startup the
     Start containers required for Integration Tests
 
     ```bash
-    source ./docker/env.sh
     docker compose up -d mysql kafka init-kafka redis-node-0 redis-node-1 redis-node-2 redis-node-3 redis-node-4 redis-node-5
     ```
 
@@ -227,7 +226,6 @@ If you want to run integration tests in a repetitive manner, you can startup the
     Start containers required for Integration Tests, including a `central-ledger` container which will be used as a proxy shell.
 
     ```bash
-    source ./docker/env.sh
     docker-compose -f docker-compose.yml -f docker-compose.integration.yml up -d kafka mysql central-ledger init-kafka redis-node-0 redis-node-1 redis-node-2 redis-node-3 redis-node-4 redis-node-5
     ```
 
@@ -244,7 +242,6 @@ If you want to run override position topic tests you can repeat the above and us
 #### For running integration tests for batch processing interactively
 - Run dependecies
 ```bash
-source ./docker/env.sh
 docker compose up -d mysql kafka init-kafka redis-node-0 redis-node-1 redis-node-2 redis-node-3 redis-node-4 redis-node-5
 npm run wait-4-docker
 ```
