+++ package.json
@@ -63,7 +63,7 @@
     "migrate:current": "npx knex migrate:currentVersion $npm_package_config_knex",
     "seed:run": "npx knex seed:run $npm_package_config_knex",
     "docker:build": "docker build --build-arg NODE_VERSION=\"$(cat .nvmrc)-alpine\" -t mojaloop/central-ledger:local .",
-    "docker:up": ". ./docker/env.sh && docker-compose -f docker-compose.yml up -d",
+    "docker:up": "docker-compose -f docker-compose.yml up -d",
     "docker:up:backend": "docker-compose up -d ml-api-adapter mysql mockserver kafka kowl temp_curl",
     "docker:up:int": "docker compose up -d kafka init-kafka objstore mysql",
     "docker:script:populateTestData": "sh ./test/util/scripts/populateTestData.sh",
@@ -119,8 +119,10 @@
     "lodash": "4.17.21",
     "moment": "2.30.1",
     "mongo-uri-builder": "^4.0.0",
+    "node-redlock": "^2.2.0",
     "parse-strings-in-object": "2.0.0",
     "rc": "1.2.8",
+    "redis-semaphore": "^5.6.2",
     "redlock": "5.0.0-beta.2",
     "require-glob": "^4.1.0"
   },
